Change Result enum underlying type from int to uint8_t for better memory efficiency

+++ include/mcutils/Result.h
@@ -27,7 +27,7 @@ namespace mc {
 /**
  * \brief The Result enum
  */
-enum class Result : int
+enum class Result : uint8_t
 {
     Success = 0,    ///< success
     Failure = 1     ///< failure
